chore: migrate all to esm
