+++ src/commands/sites/sites-list.mjs
@@ -1,6 +1,6 @@
 // @ts-check
 import { listSites } from '../../lib/api.mjs'
-import { startSpinner, stopSpinner } from '../../lib/spinner.cjs'
+import { startSpinner, stopSpinner } from '../../lib/spinner.mjs'
 import { chalk, log, logJson } from '../../utils/command-helpers.mjs'
 
 /**
