+++ src/lib/completion/constants.mjs
@@ -1,4 +1,4 @@
 // @ts-check
-import { getPathInHome } from '../settings.cjs'
+import { getPathInHome } from '../settings.mjs'
 
 export const AUTOCOMPLETION_FILE = getPathInHome(['autocompletion.json'])
