+++ src/lib/edge-functions/internal.mjs
@@ -3,7 +3,7 @@ import { readFile, stat } from 'fs/promises'
 import { dirname, join, resolve } from 'path'
 import { cwd } from 'process'
 
-import { getPathInProject } from '../settings.cjs'
+import { getPathInProject } from '../settings.mjs'
 
 import { INTERNAL_EDGE_FUNCTIONS_FOLDER } from './consts.mjs'
