+++ src/utils/rules-proxy.mjs
@@ -6,7 +6,7 @@ import cookie from 'cookie'
 import redirector from 'netlify-redirector'
 import pFilter from 'p-filter'
 
-import { fileExistsAsync } from '../lib/fs.cjs'
+import { fileExistsAsync } from '../lib/fs.mjs'
 
 import { NETLIFYDEVLOG } from './command-helpers.mjs'
 import { parseRedirects } from './redirects.mjs'
